Moved video sampler to video.py and added multiple segments sampled per path
