+++ rl_teacher/segment_sampling.py
@@ -13,7 +13,6 @@ def create_segment_q_states(segment):
 def sample_segment_from_path(path, segment_length):
     """Returns a segment sampled from a random place in a path. Returns None if the path is too short"""
     path_length = len(path['obs'])
-
     if path_length < segment_length:
         return None
 
@@ -28,53 +27,6 @@ def sample_segment_from_path(path, segment_length):
     # Add q_states
     segment['q_states'] = create_segment_q_states(segment)
     return segment
-
-class SegmentVideoRecorder(object):
-    def __init__(self, predictor, env, save_dir, n_desired_videos_per_checkpoint=1, checkpoint_interval=10):
-        self.predictor = predictor
-        self.env = env
-        self.n_desired_videos_per_checkpoint = n_desired_videos_per_checkpoint
-        self.checkpoint_interval = checkpoint_interval
-        self.save_dir = save_dir
-
-        self._counter = 0  # Internal counter of how many videos we've saved at a given iteration.
-
-    def path_callback(self, path, iteration):
-        if iteration % self.checkpoint_interval == 0:
-            if self._counter < self.n_desired_videos_per_checkpoint:
-                fname = '%s/run_%s_%s.mp4' % (self.save_dir, iteration, self._counter)
-                print("Saving video of run %s_%s to %s" % (iteration, self._counter, fname))
-                full_run = sample_segment_from_path(path, len(path['obs']))
-                write_segment_to_video(full_run, fname, self.env)
-                self._counter += 1
-        else:
-            self._counter = 0
-        self.predictor.path_callback(path, iteration)
-
-    def predict_reward(self, path):
-        return self.predictor.predict_reward(path)
-
-class RandomRolloutSegmentCollector(object):
-    def __init__(self, n_desired_segments, fps):
-        self.n_desired_segments = n_desired_segments
-        self.fps = fps
-        self.segments = []
-
-    def predict_reward(self, path):
-        epsilon = 1e-9  # Reward is unused during random rollout so we return a tiny value for each q_state
-        return np.ones(len(path["obs"])) * epsilon
-
-    def path_callback(self, path, iteration):
-        clip_length_in_seconds = 1.5
-        segment = sample_segment_from_path(path, int(clip_length_in_seconds * self.fps))
-        if segment:
-            self.segments.append(segment)
-
-        if len(self.segments) % 10 == 0 and len(self.segments) > 0:
-            print("Collected %s/%s segments" % (len(self.segments), self.n_desired_segments))
-
-        if len(self.segments) >= self.n_desired_segments:
-            raise SuccessfullyCollectedSegments()
 
 def random_action(env, ob):
     """ Pick an action by uniformly sampling the environment's action space. """
@@ -104,16 +56,20 @@ def do_rollout(env, action_function):
     return path
 
 def segments_from_rand_rollout(env, n_desired_segments, clip_length_in_seconds):
+    segment_length = int(clip_length_in_seconds * env.fps)
     segments = []
     while len(segments) < n_desired_segments:
         path = do_rollout(env, random_action)
+        # Calculate the number of segments to sample from the path
+        # Such that the probability of sampling the same part twice is fairly low.
+        segments_for_this_path = max(1, int(0.25 * len(path['obs']) / segment_length))
+        for _ in range(segments_for_this_path):
+            segment = sample_segment_from_path(path, segment_length)
+            if segment:
+                segments.append(segment)
 
-        segment = sample_segment_from_path(path, int(clip_length_in_seconds * env.fps))
-        if segment:
-            segments.append(segment)
-
-        if len(segments) % 10 == 0 and len(segments) > 0:
-            print("Collected %s/%s segments" % (len(segments), n_desired_segments))
+            if len(segments) % 10 == 0 and len(segments) > 0:
+                print("Collected %s/%s segments" % (len(segments), n_desired_segments))
 
     print("Successfully collected %s segments" % len(segments))
     return segments
